# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IIECS-101 Attendance & Invoice Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment records
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Invoices ──────────────────────────────────────────────────────────────
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Generate an invoice for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/generate-monthly:
    post:
      operationId: generateMonthlyInvoices
      tags: [invoices]
      summary: Generate monthly invoices for all students
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMonthlyInput"
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateMonthlyResult"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/pdf:
    get:
      operationId: downloadInvoicePdf
      tags: [invoices]
      summary: Download invoice PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
components:
  schemas:
...
    Payment:
      type: object
      required: [id, studentId, amount, status, createdAt]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: ["string", "null"]
        amount:
          type: number
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        referenceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string
        amount:
          type: number
          description: "Ignored by the server; fixed student fee is Rs 2000"
        description:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    PaymentUpdate:
      type: object
      properties:
        status:
          type: string
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        referenceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Invoice ──────────────────────────────────────────────────────────────
    Invoice:
      type: object
      required: [id, studentId, invoiceNumber, amount, status, createdAt]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: ["string", "null"]
        paymentId:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        amount:
          type: number
        issuedDate:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string
        paymentId:
          type: ["string", "null"]
        amount:
          type: number
          description: "Ignored by the server; fixed student fee is Rs 2000"
        dueDate:
          type: ["string", "null"]
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
    GenerateMonthlyInput:
      type: object
      properties:
        month:
          type: string
          description: "Month in YYYY-MM format. Defaults to current month."
    GenerateMonthlyResult:
      type: object
      required: [month, created, skipped]
      properties:
        month:
          type: string
        created:
          type: integer
        skipped:
          type: integer
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"