# lib/api-zod/src/generated/types/invoiceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IIECS-101 Attendance & Invoice Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceInput {
  studentId: string;
  /** @nullable */
  paymentId?: string | null;
  /** Ignored by the server; fixed student fee is Rs 2000 */
  amount?: number;
  /** @nullable */
  dueDate?: string | null;
}