# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IIECS-101 Attendance & Invoice Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  studentId: string;
  /** Ignored by the server; fixed student fee is Rs 2000 */
  amount?: number;
  description?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  notes?: string | null;
}